+++ CHANGELOG.md
@@ -12,6 +12,13 @@ Types of changes are:
 
 ## [Unreleased]
 
+## [0.28.0] - 2022-12-16
+
+### Features
+
+- `click_utils.SetOptionFromConfigCallback` that if used as a callback to `click.Parameter` sub-classes will get a value for the parameter from config if not provided on the command line.
+- `decorators.files_folders` can be specified in config.
+
 ## [0.27.0] - 2022-12-15
 
 ### Features
@@ -394,7 +401,8 @@ Commands can raise `AssertionError` exceptions to tell `delfino` some pre-condit
 
 - Initial copy of source codes.
 
-[Unreleased]: https://github.com/radeklat/delfino/compare/0.27.0...HEAD
+[Unreleased]: https://github.com/radeklat/delfino/compare/0.28.0...HEAD
+[0.28.0]: https://github.com/radeklat/delfino/compare/0.27.0...0.28.0
 [0.27.0]: https://github.com/radeklat/delfino/compare/0.26.1...0.27.0
 [0.26.1]: https://github.com/radeklat/delfino/compare/0.26.0...0.26.1
 [0.26.0]: https://github.com/radeklat/delfino/compare/0.25.0...0.26.0